+++ CHANGELOG.md
@@ -36,7 +36,7 @@
 - [ ] Well-explained sample usage of the demo in `.ipynb` format.
 - [x] Unit tests for staged pipeline operations.
 - [x] Unit tests for functional pipeline operations.
-- [ ] Use logger instead of print
+- [x] Use logger instead of print
 - [x] Specify stage name
 
 ### v0.1.4